pdf_llm/requirements.txt
flask==3.1.0
flask-cors==5.0.1
werkzeug==3.1.3
python-dotenv==1.0.0
langchain
langchain-community
langchain-core
langchain-openai
# The correct package name is langchain-text-splitters (with an 's' at the end)
langchain-text-splitters
openai==1.6.0
pydantic==2.5.3
pandas==2.2.0
PyPDF2==3.0.1
chromadb==0.4.22
gunicorn==21.2.0